refactor: enum tests into separate files
